Remove incorrect "_format_usage" method in custom HelpFormatter
"_format_usage" was overridden in the custom argparse HelpFormatter
and was incorrectly trying to capitalize the first character of the help
usage. It was actually just making the metavars of the usage line
lowercase. Overriding the "add_usage" method of the help formatter class
in order to modify the help usage prefix is a better solution anyways.