Drop "loop" A2SQueryContext arg and require keyword for all args
Having the ability to explicitly pass an event loop to the class
constructor of A2SQueryContext is dangerous. asyncio.get_event_loop()
should just be used as well and the user should be aware of the dangers
of instantiating an A2SQueryContext outside of a coroutine function
context.
See the following regarding general deprecation of explicit loop passing
arguments:
- https://aiohttp.readthedocs.io/en/stable/faq.html#why-is-creating-a-clientsession-outside-of-an-event-loop-dangerous